+++ app/src/main/java/org/breezyweather/sources/dmi/json/DmiTimeserie.kt
@@ -1,0 +1,36 @@
+/**
+ * This file is part of Breezy Weather.
+ *
+ * Breezy Weather is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation, version 3 of the License.
+ *
+ * Breezy Weather is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+ * License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Breezy Weather. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.breezyweather.sources.dmi.json
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+import org.breezyweather.common.serializer.DateSerializer
+import java.util.Date
+
+@Serializable
+data class DmiTimeserie(
+    @Serializable(DateSerializer::class) val localTimeIso: Date,
+    val temp: Float?,
+    val symbol: Int?,
+    @SerialName("precip1") val precip: Float?,
+    val windDegree: Float?,
+    val windSpeed: Float?,
+    val windGust: Float?,
+    val humidity: Float?,
+    val pressure: Float?,
+    val visibility: Float?
+)